requirements.txt
# Rive Navigator - Python Dependencies
google-adk>=1.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
vtracer>=0.6.12
